pyproject.toml
[project]
name = "agents-framework"
requires-python = ">=3.12, <3.13"
readme = "README.md"
version = "0.1.0"
authors = [
    {name = "Karol Wójciak", email = "karolwojciak@student.agh.edu.pl"},
    {name = "Łukasz Zawadzki", email = "lzawadzki@student.agh.edu.pl"},
]
dependencies = [
    "fastapi==0.116.1",
    "python-dotenv==1.1.1",
    "sqlmodel==0.0.25",
    "sqlalchemy==2.0.43",
    "asyncpg==0.30.0",
    "pydantic==2.11.7",
    "python-socketio==5.13.0",
    "uvicorn==0.35.0",
    "langchain==0.3.27",
    "langchain-openai==0.3.29",
    "alembic==1.16.4",
    "pyjwt==2.10.1",
    "python-multipart==0.0.20",
    "streamlit==1.48.0",
    "streamlit-pydantic==0.6.1rc3",
    "streamlit-flow-component==1.5.0",
    "websocket-client==1.8.0",
]

[dependency-groups]
dev = [
    "ruff==0.12.12",
    "pre-commit==4.3.0",
]
test = [
    "testcontainers==4.12.0",
    "pytest==8.4.2",
    "pytest-asyncio==1.1.0",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "B008",  # do not perform function calls in argument defaults
    "B904",  # allow raising exceptions without from e
]

[tool.setuptools]
py-modules = []

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
asyncio_default_test_loop_scope = "session"
